+++ pysys/config/descriptor.py
@@ -322,6 +322,8 @@ class TestModesConfigHelper:
 	the list of inherited modes, to project properties and also helper functions for combining multiple mode lists into 
 	one and for configuring a collection of modes as primary modes. 
 	
+	See :doc:`/pysys/UserGuide` for detailed information about what you can do with PySys modes. 
+	
 	:ivar list[dict[str,obj]] ~.inheritedModes: A list of the inherited modes, each defined by a dictionary containing a ``mode`` 
 		key and any number of additional parameters. 
 	:ivar ~.constants: A reference to `pysys.constants` which can be used to access constants such as ``IS_WINDOWS`` for 
@@ -358,7 +360,7 @@ class TestModesConfigHelper:
 		You would typically combine test-specific behaviour modes with any inherited execution environment modes like 
 		this::
 		
-			lambda modes: modes.combineModeDimensions(
+			lambda modes: modes.createModeCombinations(
 				helper.inheritedModes,
 				helper.makeAllPrimary(
 					{
@@ -377,26 +379,31 @@ class TestModesConfigHelper:
 		for m in modes: m['isPrimary'] = True
 		return modes
 
-	def combineModeDimensions(self, *dimensions):
+	def createModeCombinations(self, *dimensions):
 		"""
-		Generates a single flat mode list containing all combinations that can be generated the specified mode lists. 
+		Generates a mode list containing all the combinations from each mode list passed into the function. 
 		
-		For example::
+		For example, you could combine a list of inherited modes (defined in parent directories' ``pysysdirconfig.xml`` 
+		files), with a second dimension containing modes for each database you want to test with and a third dimension with 
+		modes for each web browser. The result would be a single (flat) list containing modes, with names and parameter 
+		dictionaries automatically merged together from each input dimension::
+				
+			lambda helper: helper.createModeCombinations(
+				helper.inheritedModes,
+				{
+					'MySQL':  {'db': 'MySQL',  'dbTimeoutSecs':60}, 
+					'SQLite': {'db': 'SQLite', 'dbTimeoutSecs':120},
+					'Mock':   {'db': 'Mock',   'dbTimeoutSecs':30},
+				},
+				
+				# can use dict or list format for each mode list, whichever is more convenient: 
+				[ 
+					{'browser':'Chrome'}, # if mode is not explicitly specified it is auto-generated from the parameter(s)
+					{'browser':'Firefox'},
+				]
+				)
 		
-			lambda helper: helper.combineModeDimensions(
-			helper.inheritedModes,
-			{
-				'MySQL':  {'db': 'MySQL',  'dbTimeoutSecs':60}, 
-				'SQLite': {'db': 'SQLite', 'dbTimeoutSecs':120},
-				'Mock':   {'db': 'Mock',   'dbTimeoutSecs':30},
-			},
-			# can use dict or list format for each mode list, whichever is more convenient: 
-			[ 
-				{'browser':'Chrome'}, # if mode is not explicitly specified it is auto-generated from the parameter(s)
-				{'browser':'Firefox'},
-			])
-		
-		would generate modes named::
+		would generate a list of modes named::
 		
 			MySQL_Chrome
 			MySQL_Firefox
@@ -408,7 +415,7 @@ class TestModesConfigHelper:
 		NB: By default the first mode in each dimension is designated a *primary* mode (one that executes by default 
 		when no ``--modes`` or ``--ci`` argument is specified), but this can be overridden by setting ``'isPrimary': True/False`` 
 		in the dict for any mode. When mode dimensions are combined, the primary modes are AND-ed together, 
-		i.e. any where *all* mode dimensions were designated primary. 
+		i.e. any where *all* mode dimensions will be designated primary. 
 		So in the above case, since MySQL and Chrome are automatically set as 
 		primary modes, the MySQL_Chrome mode would be the (only) primary mode returned from this function.
 		When using modes for different execution environments/browsers etc you probably want only 
@@ -420,7 +427,7 @@ class TestModesConfigHelper:
 		A common use case is to combine inherited modes from the parent pysysdirconfigs with a list of modes specific to 
 		this test::
 		
-			lambda helper: helper.combineModeDimensions(
+			lambda helper: helper.createModeCombinations(
 				helper.inheritedModes,
 				
 				helper.makeAllPrimary(
@@ -433,20 +440,29 @@ class TestModesConfigHelper:
 							'expectedExitStatus':'!=0', 'expectedMessage':'Server failed: Cannot specify port twice'}, 
 					}), 
 			)
+		
+		For simplicity this common case can be expressed with the more concise syntax::
+		
+			__pysys_parameterized_test_modes__ = {
+					'Usage':        {'cmd': ['--help'], 'expectedExitStatus':'==0'}, 
+					'BadPort':      {'cmd': ['--port', '-1'],  'expectedExitStatus':'!=0'}, 
+					'MissingPort':  {'cmd': [],  'expectedExitStatus':'!=0'}, 
+				}
 
-		This is a good way to use modes for the concept of parameterized subtests, since even if you don't initially have 
-		any inherited modes, if in future you add some then everything will automatically work correctly. 
-
-		NB: For efficiency reasons, don't use the ``combineModeDimensions`` method in your configuration if you are 
+		NB: For efficiency reasons, don't use the ``createModeCombinations`` method in your configuration if you are 
 		*just* using the inherited modes unchanged. 
 		
 		:param list[dict[str,obj]]|dict[str,dict[str,obj]] dimensions: Each argument passed to this function is a list of 
 			modes, each mode defined by a dict which may contain a ``mode`` key plus any number of parameters. 
 			Alternatively, each dimension can be a dict where the mode is the key and the value is a parameters dict. 
+			
 		:return: A list[dict[str,obj]] containing the flattened list of modes consisting of all combinations of the 
-			passed in. For example: ``[{"mode":"MyMode", "param1":100}, {"mode": "myMode2", "param1":200}]``. 
-			This list can be further manipulated using Python list comprehensions (e.g. to exclude certain 
+			passed in. Mode names and parameter dictionaries will be merged. 
+			For example: ``[{"mode":"MyMode", "param1":100}, {"mode": "myMode2", "param1":200}]``. 
+			The returned list can be further manipulated using Python list comprehensions (e.g. to exclude certain 
 			combinations) if desired. 
+			
+		.. versionadded:: 2.1
 		"""
 		if len(dimensions) == 1: return dimensions[0]
 		
@@ -459,7 +475,6 @@ class TestModesConfigHelper:
 				dimension = [{**{'mode':k}, **v} for k, v in dimension.items()]
 
 			for mode in dimension:
-				assert isinstance(mode, dict), 'Each mode must be a {...} dict but found unexpected object %r (%s)'%(mode, mode.__class__.__name__)
 				modeString, params = _XMLDescriptorParser.splitModeNameAndParams(mode, project=self.project)
 				current[modeString] = mode
 			# end for mode
@@ -484,7 +499,12 @@ class TestModesConfigHelper:
 							current[modeA.strip('_')+'_'+modeB.strip('_')] = params
 		return [{**{'mode':modeString}, **params} for modeString, params in current.items()]
 
-	
+	def combineModeDimensions(self, *dimensions):
+		"""
+		Old name for the `createModeCombinations` method. 
+		"""
+		return self.createModeCombinations(*dimensions)
+
 
 class TestMode(str): # subclasses string to retain compatibility for tests that don't use mode parameters
 	"""Represents a mode that a test can run in in a `TestDescriptor`, and optionally a dict of parameters that define 
@@ -540,7 +560,6 @@ class TestMode(str): # subclasses string to retain compatibility for tests that 
 	
 	def __repr__(self):
 		return self.name+str(self.__params)+('[PRIMARY]' if self.__isPrimary else '')
-			
 	
 class _XMLDescriptorParser(object):
 	'''NOT PUBLIC API - use L{DescriptorLoader._parseTestDescriptor} instead. 
@@ -883,20 +902,28 @@ class _XMLDescriptorParser(object):
 		"""
 		Returns (modename, params). Auto-generates a mode name if one is not already provided. 
 		
-		The mode dict is mutated by this method. 
+		WARNING: The mode dict is mutated by this method. 
 		"""
-		assert isinstance(mode, dict), 'Expecting mode dict but got %r'%mode
-		modeString = mode.pop('mode', None)
-		if modeString: return modeString, mode
+		
+		if isinstance(mode, TestMode): # just for the parameterized test modes use case
+			modeString = mode.name
+			params = dict(mode.params)
+			params['isPrimary'] = mode.isPrimary
+			mode = params
+		else:
+			assert isinstance(mode, dict), 'Each mode must be a {...} dict but found unexpected object %r (%s)'%(mode, mode.__class__.__name__)
 
-		# Auto-generate mode string
-		
-		assert len(mode) != 0, 'Must provide a name and/or params for every mode dictionary'
-		modeString = '_'.join(
-			'%s=%s'%(k, v) if (not isinstance(v, str) or re.match('^([-0-9.]+|true|false|)$', v, flags=re.IGNORECASE)) else v # include the key for numeric and boolean values
-			for (k,v) in mode.items() if k != 'isPrimary')
-		
-		modeString = modeString.strip('_') # avoid leading/trailing _'s
+			modeString = mode.pop('mode', None)
+			if modeString: return modeString, mode
+
+			# Auto-generate mode string
+			
+			assert len(mode) != 0, 'Must provide a name and/or params for every mode dictionary'
+			modeString = '_'.join(
+				'%s=%s'%(k, v) if (not isinstance(v, str) or re.match('^([-0-9.]+|true|false|)$', v, flags=re.IGNORECASE)) else v # include the key for numeric and boolean values
+				for (k,v) in mode.items() if k != 'isPrimary')
+			
+			modeString = modeString.strip('_') # avoid leading/trailing _'s
 
 		# Enforce consistent naming convention of initial caps
 		if project.getProperty('enforceModeCapitalization', True):
@@ -904,7 +931,24 @@ class _XMLDescriptorParser(object):
 
 		assert modeString, 'Mode name cannot be empty'
 		return modeString, mode
+
+
+	def _addParameterizedTestModes(self, base):
+		# base: list(TestMode)
+		parameterized = self.kvDict.pop('parameterized_test_modes', None)
+		if not parameterized: return base
 		
+		assert isinstance(parameterized, list) or isinstance(parameterized, dict), 'Parameterized test modes must be a list or dict of modes, but got: %r'%parameterized
+		helper = TestModesConfigHelper(
+				inheritedModes=[], 
+				project=self.project, 
+				testDir=os.path.dirname(self.file) if self.istest else None
+				)
+		return [TestMode(m.pop('mode'), isPrimary=m.pop('isPrimary', False), params=m) for m in 
+			helper.createModeCombinations(
+				base, 
+				helper.makeAllPrimary(parameterized))]
+					
 	def getModes(self):
 		text = self.kvDict.pop('modes', None)
 		modesNode = None
@@ -913,7 +957,7 @@ class _XMLDescriptorParser(object):
 			if modesNode:
 				text = self.getText(modesNode)
 			else: # if we have neither kvText text nor mode
-				return self.defaults.modes # by default we inherit
+				return self._addParameterizedTestModes(self.defaults.modes) # by default we inherit
 
 		if not text: 
 			# pre-2.0 XML approach
@@ -933,17 +977,17 @@ class _XMLDescriptorParser(object):
 			if result and not any(m.isPrimary for m in result): 
 				result[0] = TestMode(result[0], params=result[0].params, isPrimary=True)
 
-			return result
+			return self._addParameterizedTestModes(result)
 			
 		# The modern PySys 2.0+ approach with a Python eval string
 		try:
 			modesLambda = text
-			if isinstance(modesLambda, str): # eventually kvDict may be able to contain non-string values direct from Python
+			if isinstance(modesLambda, str): # kvDict may contain non-string values direct from Python
 				# use an empty namespace since if we were parsing this as real python, all import statements would be appearing later
 				modesLambda = pysys.utils.safeeval.safeEval(text.strip(), extraNamespace={}, emptyNamespace=True)
 			assert callable(modesLambda), 'Expecting callable (e.g. lambda helper: []) but got %r'%modesLambda
 			helper = TestModesConfigHelper(
-				# note that inheritedModes param dicts may be mutated (so good think we'd creating a new dict here for the default modes)
+				# note that inheritedModes param dicts may be mutated (so good thing we'd creating a new dict here for the default modes)
 				inheritedModes=[{**mode.params, **{'mode':mode, 'isPrimary':mode.isPrimary}} for mode in self.defaults.modes], 
 				project=self.project, 
 				testDir=os.path.dirname(self.file) if self.istest else None
@@ -953,8 +997,12 @@ class _XMLDescriptorParser(object):
 			if isinstance(modes, dict):
 				modes = [{**{'mode':k}, **v} for k, v in modes.items()]
 			assert isinstance(modes, list), 'Expecting a list of modes, got a %s: %r'%(modes.__class__.__name__, modes)
-			assert not modesNode or not modesNode.hasAttribute('inherit'), 'Cannot use the legacy inherit= attribute when using the modern Python eval string to define modes'
+			assert not modesNode or not modesNode.hasAttribute('inherit'), 'Cannot use the legacy inherit= attribute when using the modern Python lambda to define modes'
 			
+
+			# Add parameterized modes before validation
+			modes = self._addParameterizedTestModes(modes)
+
 			result = []
 			already = set()
 			expectedparams = None
@@ -963,7 +1011,6 @@ class _XMLDescriptorParser(object):
 				isPrimary = params.pop('isPrimary', False)
 				assert isPrimary in [True, False], 'isPrimary must be set to True or False, not %r'%isPrimary
 
-			
 				# Eliminate dodgy characters
 				badchars = re.sub('[%s]+'%pysys.launcher.MODE_CHARS,'', modeString)
 				if badchars: 
@@ -985,12 +1032,13 @@ class _XMLDescriptorParser(object):
 				for p in params:
 					assert not p.startswith('_'), 'Illegal mode parameter name - cannot start with underscore: %s'%p
 
-
 				# check that params are the same in each one to avoid mistakes
 				if expectedparams is None: expectedparams = sorted(params.keys())
 				assert sorted(params.keys()) == expectedparams, f'The same mode parameter keys must be given for every mode in the list, but found {sorted(params.keys())} parameters for "{modeString}" different to {expectedparams}'
 
 				result.append(TestMode(modeString, isPrimary=isPrimary, params=params))
+
+			result = self._addParameterizedTestModes(result)
 
 			# ensure there's at least one primary mode
 			if result and not any(m.isPrimary for m in result): 
@@ -999,8 +1047,7 @@ class _XMLDescriptorParser(object):
 			return result
 
 		except Exception as ex:
-			if not isinstance(ex, AssertionError):
-				traceback.print_exc()
+			log.debug('Invalid modes config: ', exc_info=True)
 			raise UserError("Invalid modes configuration in %s: %s"%(self.file, ex))