+++ samples/getting-started/test/correctness/MyServer_004/pysystest.py
@@ -12,28 +12,21 @@ __pysys_created__ = "1999-12-31"
 __pysys_groups__  = "myServerStartup"
 #__pysys_skipped_reason__   = "Skipped until Bug-1234 is fixed"
 
-__pysys_modes__   = lambda helper: helper.combineModeDimensions(
-			# If any inherited modes were defined at the project/parent dir level to cover different 
-			# execution environments (e.g. real db, mock db, different web browsers etc) 
-			# we would want to combine each of them with each of the different behavioural scenarios
-			# covered by this test, which we can do using combineModeDimensions:
-			
-			helper.inheritedModes,
-			
-			# Designate all of the following testing scenarios with isPrimary=True so they *all* 
-			# execute during default local test runs (without the need for --modes=ALL or --ci):
-			
-			helper.makeAllPrimary(
-				{
-					'Usage':         {'cmd': ['--help'], 
-						'expectedExitStatus':'==0', 'expectedMessage':None}, 
-					'BadPort':       {'cmd': ['--port', '-1'],  
-						'expectedExitStatus':'!=0', 'expectedMessage':'Server failed: Invalid port number specified: -1'}, 
-					'SetPortTwice':  {'cmd': ['--port', '123', '--config', helper.testDir+'/myserverconfig.json'], 
-						'expectedExitStatus':'!=0', 'expectedMessage':'Server failed: Cannot specify port twice'}, 
-				}), 
-		)
-
+# This configures a named "mode" where the entire test runs with each self.mode.params dict listed below:
+__pysys_parameterized_test_modes__  = {
+	'Usage': {
+		'cmd': ['--help'], 
+		'expectedExitStatus':'==0', 
+		'expectedMessage':None}, 
+	'BadPort': {
+		'cmd': ['--port', '-1'],  
+		'expectedExitStatus':'!=0', 
+		'expectedMessage':'Server failed: Invalid port number specified: -1'}, 
+	'SetPortTwice': {
+		'cmd': ['--port', '123', '--config', '@TEST_DIR@/myserverconfig.json'], 
+		'expectedExitStatus':'!=0', 
+		'expectedMessage':'Server failed: Cannot specify port twice'}, 
+	}
 
 import pysys
 from pysys.constants import *
@@ -42,7 +35,7 @@ class PySysTest(pysys.basetest.BaseTest):
 	def execute(self):
 		server = self.startProcess(
 			command=self.project.appHome+'/my_server.%s'%('bat' if IS_WINDOWS else 'sh'), 
-			arguments=self.mode.params['cmd'], 
+			arguments=[arg.replace('@TEST_DIR@', self.descriptor.testDir) for arg in self.mode.params['cmd']], 
 			environs=self.createEnvirons(addToExePath=os.path.dirname(PYTHON_EXE), command=PYTHON_EXE),
 			stdouterr='my_server', 
 			displayName=f'my_server<{self.mode.params["cmd"]}>',